# apps/frontend/public/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/workspaces/{wsId}/features": {
      "get": {
        "operationId": "FeaturesController_findAll",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "releaseId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assigneeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentFeatureId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isEpic",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teamId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List features in workspace",
        "tags": [
          "Features"
        ]
      },
      "post": {
        "operationId": "FeaturesController_create",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeatureDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a feature",
        "tags": [
          "Features"
        ]
      }
    },
    "/api/workspaces/{wsId}/features/{id}": {
      "get": {
        "operationId": "FeaturesController_findOne",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get feature detail",
        "tags": [
          "Features"
        ]
      },
      "patch": {
        "operationId": "FeaturesController_update",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update feature",
        "tags": [
          "Features"
        ]
      },
      "delete": {
        "operationId": "FeaturesController_remove",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete feature",
        "tags": [
          "Features"
        ]
      }
    },
    "/api/workspaces/{wsId}/features/{id}/sort": {
      "patch": {
        "operationId": "FeaturesController_updateSort",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SortFeatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update feature sort order",
        "tags": [
          "Features"
        ]
      }
    },
...
    "/api/feedback/{token}": {
      "get": {
        "operationId": "FeedbackPortalController_view",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Public feedback portal: config + feedback list with votes/themes",
        "tags": [
          "FeedbackPortal"
        ]
      },
      "post": {
        "operationId": "FeedbackPortalController_submit",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Public feedback submission (lands as Support or Idea per workspace config)",
        "tags": [
          "FeedbackPortal"
        ]
      }
    },
    "/api/feedback/{token}/captcha": {
      "post": {
        "operationId": "FeedbackPortalController_captcha",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Generate arithmetic captcha (anti-spam)",
        "tags": [
          "FeedbackPortal"
        ]
      }
    },
    "/api/feedback/{token}/vote": {
      "post": {
        "operationId": "FeedbackPortalController_vote",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalVoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Public vote on a feedback item (dedup by email)",
        "tags": [
          "FeedbackPortal"
        ]
      }
    },
    "/api/workspaces/{wsId}/feedback-portal/settings": {
      "get": {
        "operationId": "FeedbackPortalController_getSettings",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get feedback portal settings",
        "tags": [
          "FeedbackPortal"
        ]
      },
      "put": {
        "operationId": "FeedbackPortalController_updateSettings",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortalSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update feedback portal settings (enabling without token generates one)",
        "tags": [
          "FeedbackPortal"
        ]
      }
    },
    "/api/workspaces/{wsId}/feedback-portal/token": {
      "post": {
        "operationId": "FeedbackPortalController_regenerateToken",
        "parameters": [
          {
            "name": "wsId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Regenerate feedback portal token (old link invalidated)",
        "tags": [
          "FeedbackPortal"
        ]
      }
    }
  },
  "info": {
    "title": "Pinecone API",
    "description": "Aha!-style requirements management system.\n\nI10 API 治理（2026-08-18 P2）：全局前缀 /api；认证 = Bearer JWT（Web）或 API Token（pc_tok_ 前缀，SHA-256 存储）；全局限流 100 req/min/IP（Throttler）；Swagger 仅非生产暴露。",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateFeatureDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseId": {
            "type": "object",
            "nullable": true
          },
          "priority": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          },
          "assigneeName": {
            "type": "string"
          },
          "effortEstimate": {
            "type": "number"
          },
          "effortUnit": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateFeatureDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "releaseId": {
            "type": "object",
            "nullable": true
          },
          "priority": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          },
          "assigneeName": {
            "type": "string"
          },
          "effortEstimate": {
            "type": "number"
          },
          "effortUnit": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "READY_FOR_GROOMING",
              "DECOMPOSITION",
              "IN_DEVELOPING",
              "IN_VERIFICATION",
              "CLOSED"
            ],
            "description": "Feature 状态切换"
          }
        }
      },
      "SortFeatureDto": {
        "type": "object",
        "properties": {
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "sortOrder"
        ]
      },
...
      "SubmitFeedbackDto": {
        "type": "object",
        "properties": {}
      },
      "PortalVoteDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePortalSettingsDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}